Publish gosec results to Code Scanning (#84)
* fix: harden production gateway behavior

Use secure defaults and explicit configuration for upstream health checks,
rate-limit outages, ACME persistence, and release artifacts.


* fix: require admin tokens beyond loopback

Prevent IP-only access to the management API when its allowlist includes
non-loopback networks.


* fix: require secure ext authz transport

Require HTTPS for external authorization endpoints unless an operator
explicitly opts into a trusted plaintext network.


* fix: require acknowledgement for insecure backend TLS

Require an explicit configuration acknowledgement before disabling upstream
certificate verification.


* feat: add global request body limit

Apply a secure default body-size cap to proxied requests while allowing
route-specific limits for intentional upload endpoints.


* fix: reject public trusted network ranges

Prevent public catch-all CIDRs from trusting forwarding headers or exposing
operational metrics endpoints.


* fix(deps): update x/text past vulnerable release


* ci: pin the verification toolchain


* ci: publish gosec findings to code scanning


* fix: stabilize config watcher debounce test

Burst file writes without inter-write sleeps so slow CI scheduling
cannot open extra debounce windows.

* ci: pin gosec version for SARIF upload

Define GOSEC_VERSION and skip the Code Scanning upload when the SARIF
artifact was never produced.

---------

+++ internal/middleware/ratelimit_store.go
@@ -93,7 +93,7 @@ func (s *memoryStore) shardFor(key string) *memoryShard {
 		h ^= uint32(key[i])
 		h *= 16777619
 	}
-	return s.shards[h%uint32(len(s.shards))]
+	return s.shards[uint64(h)%uint64(len(s.shards))]
 }
 
 func (s *memoryStore) Check(_ context.Context, key string, limit int, window time.Duration, now time.Time) (bool, int, time.Time, error) {
